feat: Add level-up panel skeleton with selectable option buttons
PlayerLevel에 OnLevelUp 이벤트 추가, GameManager가 구독해 LevelUp 상태로 전환,
LevelUpPanel이 MenuPanelBase를 상속해 상태에 따라 자동 표시·숨김되도록 구현,
LevelUpOptionButton이 콜백으로 선택을 패널에 알리는 구조 (선택지 내용은 임시 문자열),
경험치 증가 배수 1.05 → 1.1 조정, ResultPanel 시간 계산 타입 int로 수정

+++ Assets/Scripts/Runtime/Menu/ResultPanel.cs
@@ -34,8 +34,8 @@ public class ResultPanel : MenuPanelBase
         base.Show();
 
         float time = _gameManager.PlayTime;
-        float minute = Mathf.FloorToInt(time / 60);
-        float second = Mathf.FloorToInt(time % 60);
+        int minute = Mathf.FloorToInt(time / 60);
+        int second = Mathf.FloorToInt(time % 60);
 
         _playTime.text = $"{minute}분 {second}초";
         _lastLevel.text = _playerLevel.Level.ToString();
